requirements.txt
# Python 3.11+ 호환 패키지
# naver_news_rag.py 필수 의존성

# HTTP 요청 및 HTML 파싱
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<6.0.0

# Pydantic v2 (LangChain 의존성)
pydantic>=2.0.0,<3.0.0

# LangChain Core
langchain-core>=0.1.0,<0.3.0
langchain-community>=0.0.20,<0.3.0
langchain-text-splitters>=0.0.1,<0.3.0

# LangChain OpenAI
langchain-openai>=0.0.5,<0.2.0
openai>=1.0.0,<2.0.0

# 임베딩 모델
sentence-transformers>=2.2.0,<4.0.0
torch>=2.0.0,<2.5.0
transformers>=4.35.0,<4.50.0

# Vector DB
chromadb>=0.4.22,<0.6.0

# 추가 의존성
numpy>=1.24.0,<2.0.0
tqdm>=4.65.0

# 환경변수 관리
python-dotenv>=1.0.0,<2.0.0

# WebUI (Flask)
Flask>=2.0.0,<4.0.0
